fix not responding to isready in "inner" uci loop
once a go command was received from the outer uci handling loop,
that loop actually starts the search directly, so the search code
has a second uci parsing loop to be able to respond to the "stop"
command (unless it's running in weak-elo sleep(10) mode, in which
case it takes ages until it actually processes the stop command...
but that's another issue).

this architecture is pretty messy - there should be a dedicated
thread running the uci processing loop, which then notifies the
engine threads when they have to start or stop working.