don't use chdir() to open files
the engine chdir()s all over the place - basically whenever it opens
a file. this costed me hours to figure out why no core file was written
on a crash that wasn't capturable with gdb attached.
reason was that it chdir'd to /share/rodent3 where the personality and
book files are installed, but only root has write access.

i wrote a utility function that opens a file either via absolute path,
in the hardcoded library directory, or via a specific environment variable,
and if all of the above fails, a path relative to the current working dir.